+++ back-end/src/main/java/gr/uoa/di/jete/controllers/SprintController.java
@@ -1,9 +1,11 @@
 package gr.uoa.di.jete.controllers;
 
 
+import gr.uoa.di.jete.exceptions.DeveloperNotFoundException;
 import gr.uoa.di.jete.exceptions.ProjectNotFoundException;
 import gr.uoa.di.jete.exceptions.SprintNotFoundException;
 import gr.uoa.di.jete.models.*;
+import gr.uoa.di.jete.repositories.DeveloperRepository;
 import gr.uoa.di.jete.repositories.ProjectRepository;
 import gr.uoa.di.jete.repositories.SprintRepository;
 import org.springframework.web.bind.annotation.*;
@@ -23,11 +25,13 @@ public class SprintController {
     private final SprintRepository repository;
     private final SprintModelAssembler assembler;
     private final ProjectRepository projectRep;
+    private final DeveloperRepository devRep;
 
-    SprintController(SprintRepository repository, SprintModelAssembler assembler,ProjectRepository projectRep){
+    SprintController(SprintRepository repository, SprintModelAssembler assembler, ProjectRepository projectRep, DeveloperRepository devRep){
         this.repository = repository;
         this.assembler = assembler;
         this.projectRep = projectRep;
+        this.devRep = devRep;
     }
 
     //Aggregate root
@@ -42,7 +46,7 @@ public class SprintController {
     }
     // end::get-aggregate-root[]
 
-    @PostMapping("/sprints")
+    @PostMapping("projects/sprints/create")
     Sprint newSprint(@RequestBody Sprint newSprint){
         //Search for project with given id
         projectRep.findById(newSprint.getProject_id())
@@ -83,12 +87,17 @@ public class SprintController {
         return assembler.toModel(sprint);
     }
 
-    @DeleteMapping("projects/{project_id}/sprints/{id}")
-    void deleteEpic( @PathVariable Long id,@PathVariable Long project_id){
+    @DeleteMapping("projects/{project_id}/sprints/{id}/delete/{user_id}")
+    void deleteEpic( @PathVariable Long id,@PathVariable Long project_id,@PathVariable Long user_id){
         Sprint sprint = repository.findById(new SprintId(id,project_id)) //
                 .orElseThrow(()-> new SprintNotFoundException(new SprintId(id,project_id)));
         if(sprint.getStatus()!=0L)
             throw new SprintNotFoundException(0L);
+
+        //Check that the developer requesting to delete this sprint is the product owner of the project
+        Developer developer =  devRep.findById(new DeveloperId(user_id,project_id)).orElseThrow(()->new DeveloperNotFoundException(new DeveloperId(user_id,project_id)));
+        if(developer.getRole()!=1L)
+            throw new DeveloperNotFoundException(user_id,project_id);
 
         repository.deleteAllAssigneesInSprint(id,project_id);
         repository.deleteAllTasksInSprint(id,project_id);
