+++ back-end/src/main/java/gr/uoa/di/jete/controllers/StoryController.java
@@ -1,15 +1,9 @@
 package gr.uoa.di.jete.controllers;
 
 
-import gr.uoa.di.jete.exceptions.EpicNotFoundException;
-import gr.uoa.di.jete.exceptions.ProjectNotFoundException;
-import gr.uoa.di.jete.exceptions.SprintNotFoundException;
-import gr.uoa.di.jete.exceptions.StoryNotFoundException;
+import gr.uoa.di.jete.exceptions.*;
 import gr.uoa.di.jete.models.*;
-import gr.uoa.di.jete.repositories.EpicRepository;
-import gr.uoa.di.jete.repositories.ProjectRepository;
-import gr.uoa.di.jete.repositories.SprintRepository;
-import gr.uoa.di.jete.repositories.StoryRepository;
+import gr.uoa.di.jete.repositories.*;
 import org.springframework.hateoas.CollectionModel;
 import org.springframework.hateoas.EntityModel;
 import org.springframework.web.bind.annotation.*;
@@ -29,13 +23,15 @@ public class StoryController {
     private final EpicRepository epicRep;
     private final SprintRepository sprintRep;
     private final ProjectRepository projectRep;
+    private final DeveloperRepository devRep;
 
-    StoryController(StoryRepository repository, StoryModelAssembler assembler, EpicRepository epicRep, SprintRepository sprintRep, ProjectRepository projectRep){
+    StoryController(StoryRepository repository, StoryModelAssembler assembler, EpicRepository epicRep, SprintRepository sprintRep, ProjectRepository projectRep, DeveloperRepository devRep){
         this.repository = repository;
         this.assembler = assembler;
         this.epicRep = epicRep;
         this.sprintRep = sprintRep;
         this.projectRep = projectRep;
+        this.devRep = devRep;
     }
 
     //Aggregate root
@@ -70,12 +66,15 @@ public class StoryController {
                 linkTo(methodOn(StoryController.class).getAllStoriesInEpic(project_id,epic_id)).withSelfRel());
     }
 
-    @PostMapping("/projects/{project_id}/sprints&epics/{sprint_id}&{epic_id}/stories")
+    @PostMapping("/projects/{project_id}/sprints&epics/{sprint_id}&{epic_id}/stories/create")
     Story newStory(@RequestBody Story newStory,@PathVariable Long project_id,@PathVariable Long sprint_id,@PathVariable Long epic_id){
         //Search for project with given id
         projectRep.findById(project_id).orElseThrow(()-> new ProjectNotFoundException(project_id));
         epicRep.findById(new EpicId(epic_id,project_id)).orElseThrow(()-> new EpicNotFoundException(new EpicId(epic_id,project_id)));
         sprintRep.findById(new SprintId(sprint_id,project_id)).orElseThrow(()-> new SprintNotFoundException(new SprintId(sprint_id,project_id)));
+
+        Long newId = repository.findMaxId().orElse(0L);
+        newStory.setId(newId+1L);
 
         return repository.save(newStory);
     }
@@ -89,8 +88,26 @@ public class StoryController {
         return assembler.toModel(story);
     }
 
-    @DeleteMapping("/projects/{project_id}/sprints&epics/{sprint_id}&{epic_id}/stories/{id}")
-    void deleteEpic(@PathVariable Long id,@PathVariable Long project_id,@PathVariable Long sprint_id,@PathVariable Long epic_id){
+    @PutMapping("/projects/{project_id}/sprints&epics/{sprint_id}&{epic_id}/stories/{id}/archive/{user_id}")
+    void archiveStory(@PathVariable Long id,@PathVariable Long project_id,@PathVariable Long sprint_id,@PathVariable Long epic_id,@PathVariable Long user_id){
+        //Check that the developer requesting to delete this epic is the product owner of the project
+        Developer developer =  devRep.findById(new DeveloperId(user_id,project_id)).orElseThrow(()->new DeveloperNotFoundException(new DeveloperId(user_id,project_id)));
+        if(developer.getRole()!=1L)
+            throw new DeveloperNotFoundException(user_id,project_id);
+
+        repository.archiveAllTasksInStory(id,project_id,sprint_id,epic_id);
+        repository.archiveStory(id,project_id,sprint_id,epic_id);
+    }
+
+    @DeleteMapping("/projects/{project_id}/sprints&epics/{sprint_id}&{epic_id}/stories/{id}/delete/{user_id}")
+    void deleteStory(@PathVariable Long id,@PathVariable Long project_id,@PathVariable Long sprint_id,@PathVariable Long epic_id,@PathVariable Long user_id){
+        //Check that the developer requesting to delete this epic is the product owner of the project
+        Developer developer =  devRep.findById(new DeveloperId(user_id,project_id)).orElseThrow(()->new DeveloperNotFoundException(new DeveloperId(user_id,project_id)));
+        if(developer.getRole()!=1L)
+            throw new DeveloperNotFoundException(user_id,project_id);
+
+        repository.deleteAllAssigneesInStory(id,project_id,sprint_id,epic_id);
+        repository.deleteAllTasksInStory(id,project_id,sprint_id,epic_id);
         repository.deleteById(new StoryId(id,epic_id,sprint_id,project_id));
     }
 }
